+++ src/os2datascanner/projects/admin/adminapp/models/scannerjobs/scanner_helpers.py
@@ -278,6 +278,8 @@ class ScanStatus(AbstractScanStatus):
 
         get_latest_by = "scan_tag__time"
 
+        permissions = [("export_completed_scanstatus", _("Can export history of completed scans"))]
+
     def __str__(self):
         return f"{self.scanner}: {self.start_time}"